Use restartPolicy: OnFailure
This means the same pod will retry many times if there is an error
starting up an importer, instead of a new pod being created in that
case. In theory `OnFailure` is more prone to issues with the node
breaking the build, but it keeps the dashboard output much cleaner as
there is only ever one pod per job.