feat(disk): Reduce reserved disk space
By default, new Linux filesystems are created with 5% of space reserved
for the root user. This makes sense for general cases, and especially
for root filesystems, so that non-root users cannot fill up a disk
completely.

However, for Elasticsearch data and log volumes, there is generally no
reason to essentially throw away 5% of storage capacity, especially for
the most common case with Pelias where a snapshot is loaded and then
never modified.