Keep deployments if services are not enabled
The purpose of the `enabled` flag is to control the contents of
`pelias.json`, not enable or disable the deployments.

There is little cost to a Kubernetes deployment with replica count set
to zero, so setting the `replica` value is the suggested way to turn off
the interpolation or pip service.