Extract download commands to values.yaml
This allows the download commands to be completely customized.

In order for this to work, the commands have been modified to use an
environment variable for either the download URL or download path
(depending on how many files need to be downloaded).

No functional changes to the downloads have been made.

Note that YAML formatting might be critical here, the following guide is
helpful: https://yaml-multiline.info/